Update archive page load data to tRPC query
