+++ src/lib/trpc/router/index.ts
@@ -1,4 +1,5 @@
 import { router } from '../trpc';
+import { archiveRouter } from './archive';
 import { inboxRouter } from './inbox';
 import { countRouter } from './nav';
 import { trashRouter } from './trash';
@@ -6,7 +7,8 @@ import { trashRouter } from './trash';
 export const appRouter = router({
 	inbox: inboxRouter,
 	count: countRouter,
-	trash: trashRouter
+	trash: trashRouter,
+	archive: archiveRouter
 });
 
 export type AppRouter = typeof appRouter;
